chore: replace all non-"V2" components related to the explore page with their "V2" equivalents
